feat/refactor: move thunks to the separate file; add orders state in the store
