+++ src/redux/types.ts
@@ -1,4 +1,4 @@
-import { ICardCatObject, Nullable } from './../types/types';
+import { ICardCatObject, IOrderObject, Nullable } from './../types/types';
 
 export type RequestStatusType = 'idle' | 'loading' | 'failed';
 
@@ -9,4 +9,5 @@ export interface IMainState {
   currentCatId: Nullable<number>;
   isFirstLoad: boolean;
   requestStatus: RequestStatusType;
+  orders: Array<IOrderObject>;
 }
